Basic Schedule Service
I haven't fully tested it but qtmud.manager.tick() should call to
every service, and pass on any 'events' that the service 'subscribed'
to.

All I've tested was gettinng it so the Parser service could intercept
incoming client commands and, well, parse them. Seems to work, but I'm
sure there's at least a dozen things awfully wrong in it.