Start of Environments & Movement
There is now a Mover service which listens for 'move' events.

It works against the Room, Container, and Physical Qualities:

Container - Give a thing contents, a list
Room - If a thing doesn't have contents, give it contents
    (this'll probably be fleshed out more to have code for in-built
    exits/entrances, which is why I went ahead and did it separate
    from Container.)
Physical - gives attributes for name, description, and location.

Now when a client logs in, their associated thing is given the Client
and Physical qualities, leaving them with connection information, a
name (for now a synonym for their identity), and moved into
qtmud.manager.back_room, a lazy little hack to give incoming clients
someplace to be until there's proper login.

I also added the whereami command so users can find the name of their
location.