Parser & Breaking Up Qualities
rewrote qtmud.services.parser.Parser to look for commands in a
thing's commands attribute, and for the command's functions to
live in the quality that gives them.

This meant breaking up the qualities from qualities/__init__.py into
individual files.

I also started documenting things using Sphinx markup. The configuration
files and such have been added to the repo. Going to try and build it
as our github pages after this commit.