Implemented Nametags
Nametags are a new thing-level attribute, and are used to find a
thing if you only have some names it might respond to. (For example
a client has the nametags 'client', 'player', 'thing', and their name
(if they've set one).

I also added a __setattr__ function to qtmud.Thing, so that
qtmud.qualities.Renderable. Essentially, if a thing has a
set_attr() function, the thing will use that when attr is being
set, rather than the type default.