Library Expansion
Lots of MUDs let you 'look at <thing in room>', even if that thing isn't a
real "item", something you can interact with. A cobblestone road might let
you 'look cobbles', for example, even though you can't do anything beside
look at the cobbles. Normally this requires a weird archaic syntax to work.

because of the granular nature of qualities, these fake-but-still-observable
items are easy to make, by making a new item and applying the "Renderable"
quality to it.

The downside is that this means a new thing is instanced for every lookable
thing in every room, which could cause memory problems down the line.

However, I think the extensibility and power this gives the engine is way
worth that potential cost. Normally it's a big commitment in MUD development
to move a thing from a lookable to a genuine item - normally a complete
rewrite. In this case, however, it's as simple as
lookable.add_quality(Physical).