Sender Service, Fixing Commands
A lot of commands broke when I set up the new parser, this fixes a fair
chunk of them, but certainly not all.

I also created the Sender service, which does basically what the
Renderer service does. Leaving the Renderer service for now, because it
will probably be used to format scenes (which maybe should be called
frames lol) for users.